# rest-api/public/OpenAPISpec.json
{
	"components": {
...
		"schemas": {
			"IntId": {
				"type": "integer",
				"format": "int32",
				"description": "An integer that serves as a record id. Used for primary keys and foreign keys.",
				"minimum": 1
			},
			"Artist": {
				"description": "Artist",
				"properties": {
					"ArtistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"ArtistId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MediaType": {
				"description": "Media Type",
				"properties": {
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"MediaTypeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Genre": {
				"description": "Genre",
				"properties": {
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"GenreId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Price": {
				"type": "number",
				"format": "float",
				"description": "A price value. Has a precision of 2 significant digits after the decimal point."
			},
			"AlbumResourceTrack": {
				"properties": {
					"TrackId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					},
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Composer": {
						"type": "string"
					},
					"Milliseconds": {
						"type": "number",
						"format": "double"
					},
					"Bytes": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					},
					"MediaType": {
						"$ref": "#/components/schemas/MediaType"
					},
					"Genre": {
						"$ref": "#/components/schemas/Genre"
					}
				},
				"required": [
					"TrackId",
					"Name",
					"MediaTypeId",
					"Milliseconds",
					"UnitPrice",
					"MediaType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AlbumResource": {
				"properties": {
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Title": {
						"type": "string"
					},
					"ArtistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Artist": {
						"$ref": "#/components/schemas/Artist"
					},
					"Tracks": {
						"items": {
							"$ref": "#/components/schemas/AlbumResourceTrack"
						},
						"type": "array"
					}
				},
				"required": [
					"AlbumId",
					"Title",
					"ArtistId",
					"Artist",
					"Tracks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateErrorJSON": {
				"description": "A validation error",
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Album": {
				"description": "Album",
				"properties": {
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Title": {
						"type": "string"
					},
					"ArtistId": {
						"$ref": "#/components/schemas/IntId"
					}
				},
				"required": [
					"AlbumId",
					"Title",
					"ArtistId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AlbumTrackInsertInput": {
				"description": "Album Track Insert Input",
				"properties": {
					"Name": {
						"type": "string"
					},
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Composer": {
						"type": "string"
					},
					"Milliseconds": {
						"type": "number",
						"format": "double"
					},
					"Bytes": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					}
				},
				"required": [
					"Name",
					"MediaTypeId",
					"Milliseconds",
					"UnitPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AlbumInsertInput": {
				"description": "Album Insert Input",
				"properties": {
					"Title": {
						"type": "string"
					},
					"ArtistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Tracks": {
						"items": {
							"$ref": "#/components/schemas/AlbumTrackInsertInput"
						},
						"type": "array"
					}
				},
				"required": [
					"Title",
					"ArtistId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AlbumUpdateInput": {
				"description": "Album Update Input",
				"properties": {
					"Title": {
						"type": "string"
					}
				},
				"required": [
					"Title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ArtistInsertInput": {
				"description": "Artist Insert Input",
				"properties": {
					"Name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ArtistUpdateInput": {
				"description": "Artist Update Input",
				"properties": {
					"Name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "Sample REST API built on the Chinook database, for Hasura demo",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "benoit ranque"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/albums": {
			"get": {
				"operationId": "GetAlbums",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AlbumResource"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertAlbum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Album"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AlbumInsertInput",
								"description": "Request Body"
							}
						}
					}
				}
			}
		},
		"/albums/{AlbumId}": {
			"get": {
				"operationId": "GetAlbum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AlbumResource"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "AlbumId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateAlbum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Album"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "AlbumId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AlbumUpdateInput",
								"description": "Request Body"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAlbum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Album"
								}
							}
						}
					}
				},
				"description": "Delete the resource using the resource id. Returns the deleted resource.",
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "AlbumId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
		"/artists": {
			"get": {
				"operationId": "GetArtists",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Artist"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"Artist"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertArtist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artist"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"Artist"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ArtistInsertInput",
								"description": "Request Body"
							}
						}
					}
				}
			}
		},
		"/artists/{ArtistId}": {
			"get": {
				"operationId": "GetArtist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artist"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"Artist"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "ArtistId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateArtist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artist"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"Artist"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "ArtistId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ArtistUpdateInput",
								"description": "Request Body"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArtist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artist"
								}
							}
						}
					}
				},
				"description": "Delete the resource using the resource id. Returns the deleted resource.",
				"tags": [
					"Artist"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "ArtistId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}